refactor(training): place accelerator builder in trainer utils
